fix(config): inline StarterTemplate types into index.ts (closes dep-cruiser orphan)
Inlines TemplateFile + StarterTemplate interfaces directly into
packages/config/src/index.ts and deletes packages/config/src/starter-template.ts.

Why Option (a) inline (vs rename to types.ts / runtime consumer / rule
exclusion): the file is small (28 lines, two interfaces, no runtime,
re-export-only) and the indirection earned nothing. Inlining drops the
orphan and removes a layer.

Verification:
- node scripts/check-dep-cruiser.mjs: 1 → 0 violations
- @brainst0rm/config typecheck: passes
- as-any budget: 280/285 (unchanged)
