+++ packages/config/src/index.ts
@@ -47,4 +47,25 @@ export {
   BUSINESS_MANIFEST_FILE,
   type LoadBusinessHarnessResult,
 } from "./business-loader.js";
-export { type TemplateFile, type StarterTemplate } from "./starter-template.js";
+
+/**
+ * Starter-template types shared between `@brainst0rm/cli` (the consumer)
+ * and `@brainst0rm/archetype-*` packages (the producers).
+ */
+export interface TemplateFile {
+  /** Relative path inside the harness root. */
+  path: string;
+  /** File content (UTF-8). */
+  content: string;
+}
+
+export interface StarterTemplate {
+  /** Slug used in `--template <slug>`. */
+  slug: string;
+  /** Human-readable description for `--help`. */
+  description: string;
+  /** Archetype this template targets - written into business.toml. */
+  archetype: string;
+  /** Files to materialize relative to harness root. */
+  files: TemplateFile[];
+}
